requirements.txt
pyTelegramBotAPI==4.12.0
requests==2.28.2
python-dotenv==1.0.0
psycopg2-binary==2.9.6